Hero Video Optimization
This push allows the video hero to be **loaded** using javascript only if the screen size exceeds 1000px. It was previously being loaded regardless of if it was being used or not. This should help prevent loading an unused video on mobile and increases mobile speeds.

Mobile hero remains loaded and hidden per usual.